pyproject.toml
[project]
name = "iptvtester"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.32",
  "urllib3>=2.2",
  "PyYAML>=6.0",
  "pandas>=2.2",
  "XlsxWriter>=3.2",
  "python-dateutil>=2.9",
]

[project.optional-dependencies]
http2 = ["httpx>=0.27"]

[project.scripts]
iptvtester = "iptvtester.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["iptvtester"]
